More information from fossilsearch response
When there's a combination of no fossils that are needed plus a couple
that are invalid you would only get the list of invalid fossils but no
indication that the other fossils were not needed.

I changed the code a bit so that you get both a list of things that
aren't needed and the list of things that were invalid.

Also note that in Python you can always do `if some_list` to check if
the list is empty. Less typing and it's more pythonic apparently.